chore: replace dalle with random number and timeout
