+++ examples/framesjs-starter/.env.sample
@@ -11,8 +11,5 @@ NEXT_PUBLIC_HOST="http://localhost:3000"
 
 # OPTIONAL EXAMPLE KEYS
 # Optional, if you use this example
-EXAMPLE_SLOW_REQUEST_IMGUR_CLIENTID=""
-EXAMPLE_SLOW_REQUEST_CHATGPT_API_SECRET=""
-EXAMPLE_SLOW_REQUEST_CHATGPT_ORGANIZATION_ID=""
 KV_REST_API_URL=""
 KV_REST_API_TOKEN=""
