+++ examples/framesjs-starter/app/examples/slow-request/slow-fetch/types.ts
@@ -1,0 +1,15 @@
+export type RandomNumberRequestStateValue =
+  | {
+      error: string;
+      status: "error";
+      timestamp: number;
+    }
+  | {
+      data: number;
+      status: "success";
+      timestamp: number;
+    }
+  | {
+      status: "pending";
+      timestamp: number;
+    };
